fix(asm-params): narrow block-count fields to u16 and improve param docs

+++ crates/asm/params/src/subprotocols/admin.rs
@@ -3,10 +3,10 @@ use arbitrary::Arbitrary;
 use serde::{Deserialize, Serialize};
 use strata_crypto::threshold_signature::ThresholdConfig;
 
-/// Parameters for the admnistration subprotocol, containing ThresholdConfig for each role.
+/// Parameters for the administration subprotocol, containing [`ThresholdConfig`] for each role.
 ///
 /// Design choice: Uses individual named fields rather than `Vec<(Role, ThresholdConfig)>`
-/// to ensure structural completeness - the compiler guarantees all 4 config fields are
+/// to ensure structural completeness - the compiler guarantees all config fields are
 /// provided when constructing this struct. However, it does NOT prevent logical errors
 /// like using the same config for multiple roles or mismatched role-field assignments.
 /// The benefit is avoiding missing fields at compile-time rather than runtime validation.
@@ -19,10 +19,10 @@ pub struct AdministrationSubprotoParams {
     /// ThresholdConfig for [StrataSequencerManager](Role::StrataSequencerManager).
     pub strata_sequencer_manager: ThresholdConfig,
 
-    /// The confirmation depth (CD) setting: after an update transaction receives this many
-    /// confirmations, the update is enacted automatically. During this confirmation period,
-    /// the update can still be cancelled by submitting a cancel transaction.
-    pub confirmation_depth: u32,
+    /// The confirmation depth (CD) setting, in Bitcoin blocks: after an update transaction
+    /// receives this many confirmations, the update is enacted automatically. During this
+    /// confirmation period, the update can still be cancelled by submitting a cancel transaction.
+    pub confirmation_depth: u16,
 }
 
 /// Roles with authority in the administration subprotocol.
